% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{add_genetag}
\alias{add_genetag}
\title{prepare the gene_tag ploygons}
\usage{
add_genetag(
  data,
  tag_height = NULL,
  tag_width = NULL,
  tag_angle = NULL,
  tag_rotate_angle = NULL,
  point_number = 4
)
}
\description{
prepare the gene_tag ploygons
}
