% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{add_homology_annotation}
\alias{add_homology_annotation}
\title{Add or replace a HomologyAnnotation on a SynSpecies}
\usage{
add_homology_annotation(x, homology)
}
\arguments{
\item{x}{A \code{SynSpecies} object.}

\item{homology}{A \code{HomologyAnnotation} object.}
}
\value{
An updated \code{SynSpecies} object.
}
\description{
Add or replace a HomologyAnnotation on a SynSpecies
}
