% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{add_individual}
\alias{add_individual}
\title{Add a SynIndividual to a SynSpecies object}
\usage{
add_individual(x, individual)
}
\arguments{
\item{x}{A \code{SynSpecies} object.}

\item{individual}{A \code{SynIndividual} object.}
}
\value{
An updated \code{SynSpecies} object.
}
\description{
Add a SynIndividual to a SynSpecies object
}
