% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{add_locus_set}
\alias{add_locus_set}
\alias{add_locus_set,SynSpecies,SynLocusSet-method}
\alias{add_locus_set,SynSpecies,ANY-method}
\alias{add_locus_set,ANY,SynLocusSet-method}
\alias{add_locus_set,ANY,ANY-method}
\title{Add a locus set to a SynSpecies object}
\usage{
add_locus_set(x, locus_set)

\S4method{add_locus_set}{SynSpecies,SynLocusSet}(x, locus_set)

\S4method{add_locus_set}{SynSpecies,ANY}(x, locus_set)

\S4method{add_locus_set}{ANY,SynLocusSet}(x, locus_set)

\S4method{add_locus_set}{ANY,ANY}(x, locus_set)
}
\arguments{
\item{x}{A \code{SynSpecies} object.}

\item{locus_set}{A \code{SynLocusSet} object.}
}
\value{
An updated \code{SynSpecies} object.
}
\description{
Add a locus set to a SynSpecies object
}
