% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{add_pairwise_alignment}
\alias{add_pairwise_alignment}
\title{Add a pairwise alignment to a SynSpecies object}
\usage{
add_pairwise_alignment(x, alignment)
}
\arguments{
\item{x}{A \code{SynSpecies} object.}

\item{alignment}{A \code{SynPairAlignment} object.}
}
\value{
An updated \code{SynSpecies} object.
}
\description{
Add a pairwise alignment to a SynSpecies object
}
