% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{add_transcripts_direction}
\alias{add_transcripts_direction}
\title{angle is the base angle of isosceles
pre: the transcripts dataframe , is the output of add_transcripts, require at least the end point of each transcripts. xend, yend, y_range(the height of exon) and the ID of
each transcripts
ratio is a para to describe the ratio = the length of base of isosceles triangle / (y_range)
prototyep of this function use the ratio to calculate the actual length of triangle.
potencial para r1 and r2.
be careful about the new columns with the name "x" and "y"}
\usage{
add_transcripts_direction(
  data,
  ratio = 0.25,
  angle = pi/3,
  lengthABS = 160,
  lengthPRO = NULL
)
}
\description{
angle is the base angle of isosceles
pre: the transcripts dataframe , is the output of add_transcripts, require at least the end point of each transcripts. xend, yend, y_range(the height of exon) and the ID of
each transcripts
ratio is a para to describe the ratio = the length of base of isosceles triangle / (y_range)
prototyep of this function use the ratio to calculate the actual length of triangle.
potencial para r1 and r2.
be careful about the new columns with the name "x" and "y"
}
