% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{add_transcripts_seq_line}
\alias{add_transcripts_seq_line}
\title{plot seq stick and tri-angle for indicating the direction of the gene.}
\usage{
add_transcripts_seq_line(
  data,
  transcripts_stop_proportion = NULL,
  transcripts_stop_length = 400
)
}
\description{
pass either proportion or length.
determination of transcription direction.
}
