% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynIndividual-class.R
\name{annotation_data<-}
\alias{annotation_data<-}
\title{Coerce and assign annotation data}
\usage{
annotation_data(x) <- value
}
\arguments{
\item{x}{A \code{SynIndividual} object.}

\item{value}{A \code{GenomicRanges::GRanges} object or \code{NULL}.}
}
\value{
The updated \code{SynIndividual} object.
}
\description{
Coerce and assign annotation data
}
