% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes-S7.R
\name{class_ggexon}
\alias{class_ggexon}
\title{The major ggexon object}
\usage{
class_ggexon(
  data = waiver(),
  ...,
  layers = list(),
  scales = NULL,
  guides = NULL,
  mapping = aes(),
  theme = NULL,
  coordinates = coord_cartesian(default = TRUE),
  facet = facet_null(),
  layout = NULL,
  labels = labs(),
  plot_env = parent.frame()
)
}
\description{
The major ggexon object
}
