% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes-S7.R
\name{class_mapping}
\alias{class_mapping}
\title{The mapping class}
\usage{
class_mapping(x = list(), ..., env = globalenv())
}
\arguments{
\item{x}{A list of quosures and constants.}

\item{...}{Reserved for future expansion.}

\item{env}{An environment for symbols that are not quosures or constants.}
}
\description{
The mapping class holds a list of quoted expressions
(\link[rlang:topic-quosure]{quosures}) or constants. An object is typically
constructed using the \code{\link[ggplot2:aes]{ggplot2::aes()}} function.
}
\keyword{internal}
