% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\name{clear_patches}
\alias{clear_patches}
\title{Clear annotation patches}
\usage{
clear_patches(x, annotation = NULL)
}
\arguments{
\item{x}{A \code{SynFeatureAnnotation} or \code{SynIndividual} object.}

\item{annotation}{Optional feature-annotation layer name when \code{x} is a
\code{SynIndividual}.}
}
\value{
The updated object.
}
\description{
Clear annotation patches
}
