% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-genomictree.R
\docType{data}
\name{FacetGenomicTree}
\alias{FacetGenomicTree}
\title{ggproto backend for \code{facet_genomictree()}}
\format{
An object of class \code{FacetGenomicTree} (inherits from \code{FacetGenomics}, \code{FacetWrap}, \code{Facet}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
FacetGenomicTree
}
\description{
ggproto backend for \code{facet_genomictree()}
}
\keyword{datasets}
