% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{filter_pairwise_alignment}
\alias{filter_pairwise_alignment}
\title{Filter a pairwise alignment by minimum PAF alignment length}
\usage{
filter_pairwise_alignment(x, filter = 200, alignment = NULL)
}
\arguments{
\item{x}{A \code{SynSpecies} or \code{SynPairAlignment} object.}

\item{filter}{Minimum \code{alen} to keep.}

\item{alignment}{Optional alignment name when \code{x} is a \code{SynSpecies}.}
}
\value{
A filtered PAF-like \code{data.frame}.
}
\description{
Filter a pairwise alignment by minimum PAF alignment length
}
