% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GADclass.R
\docType{class}
\name{GAD-class}
\alias{GAD-class}
\title{GADclass (Genomic Associate Data class) basing on S4 OO system.
Author: Liu Dongyao
E-mail: dongyao@connect.hku.hk
Date: 2025-1-21
Version:1.0v
Class definition}
\description{
GADclass (Genomic Associate Data class) basing on S4 OO system.
Author: Liu Dongyao
E-mail: dongyao@connect.hku.hk
Date: 2025-1-21
Version:1.0v
Class definition
}
