% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-exon.R
\docType{data}
\name{GeomExon}
\alias{GeomExon}
\title{Geom implementation for exon features}
\format{
An object of class \code{GeomExon} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 9.
}
\usage{
GeomExon
}
\description{
Internal ggproto used by \code{\link[=geom_exon]{geom_exon()}}.
}
\keyword{datasets}
