% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynIndividual-class.R
\name{get_annotation}
\alias{get_annotation}
\title{Retrieve an annotation layer from a SynIndividual}
\usage{
get_annotation(x, name = NULL)
}
\arguments{
\item{x}{A \code{SynIndividual} object.}

\item{name}{Annotation layer name. Defaults to the active annotation.}
}
\value{
A \code{SynAnnotation} object.
}
\description{
Retrieve an annotation layer from a SynIndividual
}
