% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggexon-package.R
\docType{package}
\name{ggexon-package}
\alias{ggexon-package}
\title{ggexon: Harmonizing Visualization of Genomic Synteny Across Different Genomes}
\description{
visulize genomic tracks with associate meta information.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dongyaoliu.github.io/ggexon/}
  \item \url{https://github.com/DongyaoLiu/ggexon}
  \item Report bugs at \url{https://github.com/DongyaoLiu/ggexon/issues}
}

}
\author{
\strong{Maintainer}: Dongyao LIU \email{dongyao@connect.hku.hk} (\href{https://orcid.org/0000-0002-8066-1487}{ORCID})

}
\keyword{internal}
