% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.R
\name{ggexon_gtable}
\alias{ggexon_gtable}
\title{Render a built ggexon plot to a gtable}
\usage{
ggexon_gtable(data)
}
\arguments{
\item{data}{A built \code{ggexon} plot object.}
}
\value{
A \link[gtable:gtable]{gtable::gtable} object ready to draw with \code{\link[grid:grid.draw]{grid::grid.draw()}}.
}
\description{
Render a built ggexon plot to a gtable
}
