% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all-classes-S7.R
\docType{data}
\name{ggplot2_class_reexports}
\alias{ggplot2_class_reexports}
\alias{class_gg}
\alias{class_ggproto}
\alias{class_scale}
\alias{class_guides}
\alias{class_guide}
\alias{class_coord}
\alias{class_facet}
\alias{class_layer}
\alias{class_layout}
\alias{class_scales_list}
\alias{class_S3_gg}
\alias{class_rel}
\alias{class_zero_grob}
\alias{class_waiver}
\alias{class_derive}
\alias{class_theme}
\alias{class_labels}
\alias{class_mapping}
\alias{class_ggplot}
\title{ggplot2 Class Objects Re-exported by ggexon}
\format{
An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.

An object of class \code{S7_S3_class} of length 3.
}
\usage{
class_gg()

class_ggproto

class_scale

class_guides

class_guide

class_coord

class_facet

class_layer

class_layout

class_scales_list

class_S3_gg

class_rel

class_zero_grob

class_waiver

class_derive

class_theme(elements = list(), ..., complete = FALSE, validate = TRUE)

class_labels(labels = list(), ...)

class_mapping(x = list(), ..., env = globalenv())

class_ggplot(
  data = waiver(),
  ...,
  layers = list(),
  scales = NULL,
  guides = NULL,
  mapping = aes(),
  theme = NULL,
  coordinates = coord_cartesian(default = TRUE),
  facet = facet_null(),
  layout = NULL,
  labels = labs(),
  meta = list(),
  plot_env = parent.frame()
)
}
\description{
ggexon re-exports ggplot2's S7 and ggproto class objects so downstream code
can target the exact same class definitions without vendoring local copies.
}
\keyword{datasets}
\keyword{internal}
