% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homology-annotation.R
\name{HomologyAnnotation}
\alias{HomologyAnnotation}
\title{Constructor for HomologyAnnotation}
\usage{
HomologyAnnotation(
  name,
  reference_species,
  query_species,
  homology_table,
  source_file = "<homology>",
  metadata = list()
)
}
\arguments{
\item{name}{Short unique label for the homology annotation layer.}

\item{reference_species}{Name of the reference (center) species.}

\item{query_species}{Name of the query species.}

\item{homology_table}{A data frame with at minimum \code{query_gene} and
\code{reference_gene} columns.}

\item{source_file}{Optional path to the source BLAST file.}

\item{metadata}{Optional metadata list.}
}
\value{
A \code{HomologyAnnotation} object.
}
\description{
Constructor for HomologyAnnotation
}
