% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggexon.R
\name{is_ggexon}
\alias{is_ggexon}
\title{Test whether an object is a ggexon plot}
\usage{
is_ggexon(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} when \code{x} is a \code{ggexon} plot object, otherwise \code{FALSE}.
}
\description{
Test whether an object is a ggexon plot
}
