% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.R
\name{layer_grob}
\alias{layer_grob}
\title{Extract a built grob for a layer}
\usage{
layer_grob(plot = get_last_plot(), i = 1L)
}
\arguments{
\item{plot}{A built ggplot or ggexon object.}

\item{i}{Layer index.}
}
\description{
\code{layer_grob} is an alias of \code{get_layer_grob()}.
}
