% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\name{list_patches}
\alias{list_patches}
\title{List annotation patches}
\usage{
list_patches(x, annotation = NULL)
}
\arguments{
\item{x}{A \code{SynFeatureAnnotation} or \code{SynIndividual} object.}

\item{annotation}{Optional feature-annotation layer name when \code{x} is a
\code{SynIndividual}.}
}
\value{
A named list of \code{SynAnnotationPatch} objects.
}
\description{
List annotation patches
}
