% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{locus_sets}
\alias{locus_sets}
\alias{locus_sets,SynSpecies-method}
\title{List locus sets attached to a SynSpecies}
\usage{
locus_sets(x)

\S4method{locus_sets}{SynSpecies}(x)
}
\arguments{
\item{x}{A \code{SynSpecies} object.}
}
\value{
A named list of \code{SynLocusSet} objects.
}
\description{
List locus sets attached to a SynSpecies
}
