% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus-set.R
\name{locus_table}
\alias{locus_table}
\alias{locus_table,SynLocusSet-method}
\title{Retrieve the locus table from a SynLocusSet}
\usage{
locus_table(x)

\S4method{locus_table}{SynLocusSet}(x)
}
\arguments{
\item{x}{A \code{SynLocusSet} object.}
}
\value{
A data frame with one row per locus window.
}
\description{
Retrieve the locus table from a SynLocusSet
}
