% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein-mutation-annotation.R
\name{protein_mutation_data}
\alias{protein_mutation_data}
\title{Access protein-mutation data}
\usage{
protein_mutation_data(x)
}
\arguments{
\item{x}{A \code{SynProteinMutationAnnotation} object.}
}
\value{
A normalized mutation \code{data.frame}.
}
\description{
Access protein-mutation data.
}
