% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig-annotation.R
\name{query_annotation}
\alias{query_annotation}
\alias{query_annotation,SynBigWigAnnotation-method}
\title{Query a region-backed annotation}
\usage{
query_annotation(x, region, ...)

\S4method{query_annotation}{SynBigWigAnnotation}(x, region, ...)
}
\arguments{
\item{x}{A region-backed \code{SynAnnotation} object.}

\item{region}{A length-one \code{GenomicRanges::GRanges} query.}

\item{...}{Additional arguments passed to a class method.}
}
\value{
A class-specific region query result.
}
\description{
Query a region-backed annotation
}
