% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\name{query_domains}
\alias{query_domains}
\title{Query protein-domain annotations}
\usage{
query_domains(x, ids = NULL, domains = NULL)
}
\arguments{
\item{x}{A \code{SynProteinDomainAnnotation} object.}

\item{ids}{Optional identifiers to match against \code{keytype}.}

\item{domains}{Optional domain names to filter.}
}
\value{
A \code{S4Vectors::DataFrame} with matching domain records.
}
\description{
Query protein-domain annotations
}
