% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\name{query_signal}
\alias{query_signal}
\title{Query signal from a SynBigWigAnnotation}
\usage{
query_signal(x, chr, start, end)
}
\arguments{
\item{x}{A \code{SynBigWigAnnotation} object.}

\item{chr}{Chromosome name.}

\item{start}{Start coordinate.}

\item{end}{End coordinate.}
}
\value{
A \code{GRanges} object with the overlapping signal records.
}
\description{
Query signal from a SynBigWigAnnotation
}
