% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homology-annotation.R
\name{query_species}
\alias{query_species}
\title{Retrieve the query species from a HomologyAnnotation}
\usage{
query_species(x)
}
\arguments{
\item{x}{A \code{HomologyAnnotation} object.}
}
\value{
A scalar character value.
}
\description{
Retrieve the query species from a HomologyAnnotation
}
