% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-null.R, R/utilities-grid.R, R/utilities.R
\name{zeroGrob}
\alias{zeroGrob}
\alias{reexports}
\alias{unit}
\alias{arrow}
\alias{alpha}
\title{Objects imported from other packages}
\usage{
zeroGrob()

unit(x, units, data = NULL)

arrow(angle = 30, length = unit(0.25, "inches"), ends = "last", type = "open")

alpha(colour, alpha = NA)
}
\description{
These helpers are re-exported so ggexon users can access them without
attaching the upstream package explicitly.
}
\examples{
ggplot(mpg, aes(displ, hwy)) +
  geom_point(alpha = 0.5, colour = "blue")

ggplot(mpg, aes(displ, hwy)) +
  geom_point(colour = alpha("blue", 0.5))
}
\keyword{internal}
