% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{seq_add_y}
\alias{seq_add_y}
\title{calculate the exon block y location
ymin and ymax is for draw the exon part}
\usage{
seq_add_y(
  data,
  track_proportion,
  exon_proportion,
  blank_proportion,
  exon_height,
  sandwich_ratio
)
}
\description{
calculate the exon block y location
ymin and ymax is for draw the exon part
}
