% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{seq_track}
\alias{seq_track}
\title{calculate the start and end of the sequencing track}
\usage{
seq_track(data, start = NULL, end = NULL)
}
\description{
calculate the start and end of the sequencing track
}
