% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynIndividual-class.R
\name{set_active_annotation}
\alias{set_active_annotation}
\title{Set the active annotation layer on a SynIndividual}
\usage{
set_active_annotation(x, name)
}
\arguments{
\item{x}{A \code{SynIndividual} object.}

\item{name}{Annotation layer name.}
}
\value{
An updated \code{SynIndividual} object.
}
\description{
Set the active annotation layer on a SynIndividual
}
