% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{store_chain_layout}
\alias{store_chain_layout}
\title{Compute and store the ggexon chain layout on a \code{SynSpecies}}
\usage{
store_chain_layout(
  x,
  vars = ggplot2::vars(track),
  free = list(x = FALSE, y = FALSE)
)
}
\arguments{
\item{x}{A \code{SynSpecies} object.}

\item{vars}{Facet vars. Defaults to \code{ggplot2::vars(track)}.}

\item{free}{List with logical \code{x} and \code{y} entries controlling scale grouping.}
}
\value{
The updated \code{SynSpecies} object.
}
\description{
Compute and store the ggexon chain layout on a \code{SynSpecies}
}
