% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{subset_pairwise_alignment}
\alias{subset_pairwise_alignment}
\alias{subset_pairwise_alignment,SynPairAlignment-method}
\alias{subset_pairwise_alignment,SynSpecies-method}
\title{Subset a pairwise alignment by query/target regions}
\usage{
subset_pairwise_alignment(x, subset, alignment = NULL)

\S4method{subset_pairwise_alignment}{SynPairAlignment}(x, subset, alignment = NULL)

\S4method{subset_pairwise_alignment}{SynSpecies}(x, subset, alignment = NULL)
}
\arguments{
\item{x}{A \code{SynSpecies} or \code{SynPairAlignment} object.}

\item{subset}{Named character vector/list with one region per species for
one or both alignment partners, e.g.
\code{c(XZ1516 = "RagTag_V:21550000-21680000", N2 = "V:20450000-20451000")},
\code{c(XZ1516 = "RagTag_V:21550000-21680000")}, or
\code{c(XZ1516 = "RagTag_V")}.}

\item{alignment}{Optional alignment name when \code{x} is a \code{SynSpecies}.}
}
\value{
An updated \code{SynPairAlignment} or \code{SynSpecies} object.
}
\description{
Subset a pairwise alignment by query/target regions
}
\details{
This is an S4 generic that dispatches on the class of \code{x}.
}
\examples{
paf_path <- system.file(
  "extdata", "cd44_pairwise_ensembl116", "cd44_lastz.paf",
  package = "ggexon"
)
pair <- SynPairAlignment(
  name = "mouse_vs_human",
  query_individual = "mouse",
  target_individual = "human",
  file = paf_path
)
pair <- subset_pairwise_alignment(pair, subset = c(mouse = "mouse"))

}
