% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\docType{class}
\name{SynBigWigAnnotation-class}
\alias{SynBigWigAnnotation-class}
\title{SynBigWigAnnotation class}
\description{
\code{SynBigWigAnnotation} stores a region-queryable signal layer backed by a
BigWig file.
}
\section{Slots}{

\describe{
\item{\code{data_format}}{Signal file format label. Currently \code{"bigwig"}.}

\item{\code{signal}}{Optional cached parsed signal data.}

\item{\code{seqinfo}}{Optional cached \code{Seqinfo} describing available sequences.}

\item{\code{window_cache}}{Cache of previously queried genomic windows.}
}}

