% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\name{SynBigWigAnnotation}
\alias{SynBigWigAnnotation}
\title{Constructor for SynBigWigAnnotation}
\usage{
SynBigWigAnnotation(name, bigwig_file, metadata = list(), lazy = TRUE)
}
\arguments{
\item{name}{Short unique label for the annotation layer.}

\item{bigwig_file}{Path to the BigWig file.}

\item{metadata}{Optional metadata list.}

\item{lazy}{Logical; defaults to \code{TRUE} for windowed loading.}
}
\value{
A \code{SynBigWigAnnotation} object.
}
\description{
Constructor for SynBigWigAnnotation
}
