% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\docType{class}
\name{SynIndAnnotation-class}
\alias{SynIndAnnotation-class}
\title{SynIndAnnotation class}
\description{
\code{SynIndAnnotation} is an abstract subtype of \code{SynAnnotation} for annotation
layers attached to a \code{SynIndividual}.
}
