% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\docType{class}
\name{SynSpeAnnotation-class}
\alias{SynSpeAnnotation-class}
\title{SynSpeAnnotation class}
\description{
\code{SynSpeAnnotation} is an abstract subtype of \code{SynAnnotation} for annotation
layers attached to a \code{SynSpecies}.
}
