% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{transcripts_track}
\alias{transcripts_track}
\title{calculate the transcripts lines coordinates.
tmp_1 extract the joint top of the angle}
\usage{
transcripts_track(
  data,
  transcript_data = NULL,
  between_margin = NULL,
  Track_Position = "below",
  joint_x = 0.5,
  joint_y = 0.2,
  transcripts_y = 0.1
)
}
\description{
calculate the transcripts lines coordinates.
tmp_1 extract the joint top of the angle
}
