% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{warn_dots_empty}
\alias{warn_dots_empty}
\title{utilities for new S7 class definition}
\usage{
warn_dots_empty(env = caller_env(), call = caller_env())
}
\description{
utilities for new S7 class definition
}
