% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{filter_pairwise_alignment}
\alias{filter_pairwise_alignment}
\title{Filter a pairwise alignment by minimum PAF alignment length}
\usage{
filter_pairwise_alignment(x, filter = 200, alignment = NULL)
}
\arguments{
\item{x}{A \code{SynSpecies} or \code{SynPairAlignment} object.}

\item{filter}{Minimum \code{alen} to keep.}

\item{alignment}{Optional alignment name when \code{x} is a \code{SynSpecies}.}
}
\value{
An updated \code{SynPairAlignment} or \code{SynSpecies} object.
}
\description{
Filter a pairwise alignment by minimum PAF alignment length
}
\details{
This is an S4 generic that dispatches on the class of \code{x}.
}
\examples{
paf_path <- system.file("extdata", "V_alginment.paf", package = "ggexon")
pair <- SynPairAlignment(
  name = "XZ1516_vs_N2",
  query_individual = "XZ1516",
  target_individual = "N2",
  file = paf_path
)
pair <- filter_pairwise_alignment(pair, filter = 200)
}
