% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-coverage.R
\docType{data}
\name{GeomCoverage}
\alias{GeomCoverage}
\title{Coverage geom ggproto}
\description{
\code{GeomCoverage} draws coverage rectangles derived from BigWig signal over a
selected genomic window.
}
\keyword{datasets}
