% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_coverage.R
\docType{data}
\name{GeomCoverage}
\alias{GeomCoverage}
\title{designed for plot coverage data.
assess the validation of annotation_type
extract the subset coordinates.
filter base on the subset region.
select transcript data.
extract the bigwig data.
deal with more than one bigwig data.
one transcript, more coverage track.
one transcript, one coverage track.
specify the multi-track multi-coverage by named list.
single track single bigwig
single track multiple bigwig
multiple track multiple bigwig files.
do y_threshold.
translation of coordinates of CoverageTable
do bigwig track normalization.
processing the coverage data and merging with track information
extract coverage based on subset genomic region.
extract base on specific region.
CoverageTable$PANLE = 1
print(CoverageTable)
all the munipulation of
./R/utilities.R:empty <- function(df) will exam the data at rendering
layout$panel_params[[data$PANEL\link{1}]] this will check the PANEL col}
\format{
An object of class \code{GeomCoverage} (inherits from \code{GeomRect}, \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomCoverage
}
\description{
designed for plot coverage data.
assess the validation of annotation_type
extract the subset coordinates.
filter base on the subset region.
select transcript data.
extract the bigwig data.
deal with more than one bigwig data.
one transcript, more coverage track.
one transcript, one coverage track.
specify the multi-track multi-coverage by named list.
single track single bigwig
single track multiple bigwig
multiple track multiple bigwig files.
do y_threshold.
translation of coordinates of CoverageTable
do bigwig track normalization.
processing the coverage data and merging with track information
extract coverage based on subset genomic region.
extract base on specific region.
CoverageTable$PANLE = 1
print(CoverageTable)
all the munipulation of
./R/utilities.R:empty <- function(df) will exam the data at rendering
layout$panel_params[[data$PANEL\link{1}]] this will check the PANEL col
}
\keyword{datasets}
