% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-coverageMUT.R
\docType{data}
\name{GeomCoverageMUT}
\alias{GeomCoverageMUT}
\title{Coverage-by-mutation geom ggproto}
\description{
\code{GeomCoverageMUT} draws coverage rectangles derived from BigWig signal and
optionally rescales them by mutation frequencies supplied in a VCF summary
table.
}
\keyword{datasets}
