% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_gene.R
\docType{data}
\name{GeomGene}
\alias{GeomGene}
\title{v=1.0
author = "LIU Dongyao"
LabWebsite: "www.zhenglabhku.org"
E-mail = "dongyao@connect.hku.hk"
Description: designed to draw gene without exon information. An arrow.}
\format{
An object of class \code{GeomGene} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomGene
}
\description{
v=1.0
author = "LIU Dongyao"
LabWebsite: "www.zhenglabhku.org"
E-mail = "dongyao@connect.hku.hk"
Description: designed to draw gene without exon information. An arrow.
}
\keyword{datasets}
