% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-genebox.R
\docType{data}
\name{GeomGeneBox}
\alias{GeomGeneBox}
\title{Gene-box geom}
\format{
An object of class \code{GeomGeneBox} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 11.
}
\usage{
GeomGeneBox
}
\description{
Public ggproto implementing \code{\link[=geom_genebox]{geom_genebox()}}.
}
\keyword{datasets}
