% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-genetag.R
\docType{data}
\name{GeomGeneTag}
\alias{GeomGeneTag}
\title{construct the function of to generate the
need one more or two parameter to control distance between the tags.
or automatically compute form the gene number and track length.}
\format{
An object of class \code{GeomGeneTag} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 8.
}
\usage{
GeomGeneTag
}
\description{
construct the function of to generate the
need one more or two parameter to control distance between the tags.
or automatically compute form the gene number and track length.
}
\keyword{datasets}
