% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R, R/SynIndividual-class.R,
%   R/homology-annotation.R, R/SynSpecies-class.R, R/ggexon-package.R
\name{show,SynAnnotation-method}
\alias{show,SynAnnotation-method}
\alias{show,SynIndividual-method}
\alias{show,HomologyAnnotation-method}
\alias{show,SynSpecies-method}
\alias{show,SynLayout-method}
\alias{ggexon-show}
\title{Display ggexon S4 objects}
\usage{
\S4method{show}{SynAnnotation}(object)

\S4method{show}{SynIndividual}(object)

\S4method{show}{HomologyAnnotation}(object)

\S4method{show}{SynSpecies}(object)

\S4method{show}{SynLayout}(object)
}
\arguments{
\item{object}{A \code{ggexon} S4 object.}
}
\value{
\code{object}, invisibly; called for the side effect of printing a summary.
}
\description{
Compact \code{show()} methods that print a one-screen summary of the major
\code{ggexon} S4 objects (\code{SynSpecies}, \code{SynIndividual}, \code{SynLayout},
\code{SynAnnotation}, and \code{HomologyAnnotation}).
}
\keyword{internal}
