% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homology-annotation.R
\name{homology_table}
\alias{homology_table}
\title{Retrieve the homology table from a HomologyAnnotation}
\usage{
homology_table(x)
}
\arguments{
\item{x}{A \code{HomologyAnnotation} object.}
}
\value{
A data frame with \code{query_gene} and \code{reference_gene} columns.
}
\description{
Retrieve the homology table from a HomologyAnnotation
}
