% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigwig-annotation.R
\name{query_signal}
\alias{query_signal}
\title{Query BigWig signal}
\usage{
query_signal(x, chr, start, end, ...)
}
\arguments{
\item{x}{A \code{SynBigWigAnnotation} object.}

\item{chr}{Chromosome name.}

\item{start}{Start coordinate.}

\item{end}{End coordinate.}

\item{...}{Additional arguments passed to \code{\link[=query_annotation]{query_annotation()}}.}
}
\value{
A \code{GenomicRanges::GRanges} object with overlapping signal records.
}
\description{
Query BigWig signal
}
