% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{species_tree<-}
\alias{species_tree<-}
\alias{species_tree<-,SynSpecies-method}
\alias{species_tree_plot<-}
\alias{species_tree_plot<-,SynSpecies-method}
\title{Store tree objects on a \code{SynSpecies}}
\usage{
species_tree(x) <- value

\S4method{species_tree}{SynSpecies}(x) <- value

species_tree_plot(x) <- value

\S4method{species_tree_plot}{SynSpecies}(x) <- value
}
\arguments{
\item{x}{A \code{SynSpecies} object.}

\item{value}{A tree object, a rectangular \code{ggtree} plot, or \code{NULL}.}
}
\value{
The updated \code{SynSpecies} object.
}
\description{
Store tree objects on a \code{SynSpecies}
}
