% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{subset_species}
\alias{subset_species}
\alias{subset_species,SynSpecies-method}
\title{Subset one or more individuals in a \code{SynSpecies} by coordinate string}
\usage{
subset_species(x, coords, annotations = c("all_feature", "active"))

\S4method{subset_species}{SynSpecies}(x, coords, annotations = c("all_feature", "active"))
}
\arguments{
\item{x}{A \code{SynSpecies} object.}

\item{coords}{One or more coordinate strings in the form
\code{"species#seqname:start-end"}. This can be a single string, a character
vector, or a list of strings.}

\item{annotations}{One of \code{"all_feature"} or \code{"active"}. Passed through to
\code{\link[=subset_individual]{subset_individual()}}.}
}
\value{
A \code{SynSpecies} object.
}
\description{
Returns a new \code{\link{SynSpecies}} in which selected individuals have their
feature annotation layers trimmed according to coordinate strings such as
\code{"XZ1516#V_RagTag:21559983-21620009"}. Individuals not listed in \code{coords}
are left unchanged. Any stored \code{\link{SynLayout}} is cleared because the panel
metadata may no longer match the new subsetted windows.
}
\details{
This is an S4 generic that dispatches on the class of \code{x}.
}
\examples{
ann_path <- system.file(
  "extdata",
  "compact_synspecies",
  "caenorhabditis_XZ1516.gff3",
  package = "ggexon"
)
ind <- SynIndividual(
  annotation_file = ann_path,
  genome_file = genome_waiver(),
  id = "XZ1516"
) |>
  load_annotation()
gr <- annotation_data(ind)
coords <- paste0(
  "XZ1516#",
  as.character(GenomeInfoDb::seqnames(gr))[[1L]],
  ":",
  IRanges::start(gr)[[1L]],
  "-",
  IRanges::end(gr)[[1L]]
)

sp <- SynSpecies(name = "worms") |> add_individual(ind)
sp_window <- subset_species(sp, coords = coords)

}
