% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\docType{class}
\name{SynGenomeAnnotation-class}
\alias{SynGenomeAnnotation-class}
\title{SynGenomeAnnotation class}
\description{
\code{SynGenomeAnnotation} is an abstract subtype of \code{SynIndAnnotation} for
annotations indexed on genomic coordinates.
}
