% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynSpecies-class.R
\name{SynMultiAlignment}
\alias{SynMultiAlignment}
\title{Constructor for SynMultiAlignment}
\usage{
SynMultiAlignment(name, individuals, file, metadata = list())
}
\arguments{
\item{name}{Alignment label.}

\item{individuals}{Character vector of included individuals.}

\item{file}{Path to the MAF file.}

\item{metadata}{Optional metadata list.}
}
\value{
A \code{SynMultiAlignment} object.
}
\description{
Constructor for SynMultiAlignment
}
