% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\docType{class}
\name{SynProteinAnnotation-class}
\alias{SynProteinAnnotation-class}
\title{SynProteinAnnotation class}
\description{
\code{SynProteinAnnotation} is an abstract subtype of \code{SynIndAnnotation} for
annotations indexed on protein coordinates.
}
