% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SynAnnotation-class.R
\docType{class}
\name{SynVCFAnnotation-class}
\alias{SynVCFAnnotation-class}
\title{SynVCFAnnotation class}
\description{
\code{SynVCFAnnotation} stores a region-queryable variant layer backed by a VCF or
BCF file.
}
\section{Slots}{

\describe{
\item{\code{data_format}}{Variant file format label. Currently \code{"vcf"}.}

\item{\code{variants}}{Optional cached parsed variant data.}

\item{\code{index_file}}{Optional index file path for random access.}

\item{\code{genome_build}}{Optional genome-build label.}

\item{\code{region_cache}}{Cache of previously queried genomic windows.}
}}

